refactor(bb): removed powers of eta in lookup and auxiliary relations (#4695)
We previously were using powers of the plookup "eta" challenge in our
lookup and auxiliary relations, to construct random linear combinations.

With protogalaxy, this becomes inefficient as each challenge is
represented by a degree-1 Univariate polynomial. i.e. eta^3 is degree-3,
while a completely independent challenge is degree-1.

Usages of eta^2 and eta^3 have been replaced with new challenges eta_two
and eta_three

Now that the `CircuitChecker` uses relations, we update these as
appropriately but also have to change `stdlib_plonk_recursion_tests` to
construct and verify full proof as the widgets do not reflect changes on
`eta`

---------

Co-authored-by: maramihali <mara@aztecprotocol.com>