chore: Improve `compute_note_hash_and_nullifier` autogeneration and `NoteProcessor` warnings (#5838)
Closes https://github.com/AztecProtocol/aztec-packages/issues/5669,
https://github.com/AztecProtocol/aztec-packages/issues/5670,
https://github.com/AztecProtocol/aztec-packages/issues/4649

Correctly determines the signature of the autogenerated
`compute_note_hash_and_nullifier` by looking up the serialized content
size of notes used in the contract. Furthermore, that langth is compared
against MAX_NOTE_FIELDS_LEN and a compile time error is emited if the
user attempts to use a note that is bigger than currently supported.
Finally changed the `NoteProcessor` warns to be errors (even thought
this particular one shouldn't get that far anymore!)