chore: add simple `bb` installer script (#6376)
As of #6320, we're now expecting users to maintain an installation of
`bb` and update its version as required.

To help with this, this PR adds an ultra dumb `bbup` script fashioned
from `noirup` which only allows installing released versions of `bb`.
This means that users can run `bbup -v 0.38.0` to install 0.38.0 of
`bb`.

Happy to make any changes as requested.