fix: add cbind declarations for new methods to fix autogen (#6622)
Some declarations for new cbind methods were missing which was causing
the autogen bindings in index.ts not to be produced.

I guess these aren't being tested? Update: seems like most are tested
and it was working because the CI does not re-autogen the bindings so if
they've been manually implemented it works?