fix: Biggroup batch mul handles collisions (#6780)
The following PR adds an edgecase handling mode to biggroup batch
multiplication. In this mode the points are randomised in such a way as
avoid weird interactions (doublings and point at infinity cases).
It enables using tables for RecurisveMergeVerifier and Recursive
Verifier for Protogalaxy on ultra (it also halves the gatecount from 4.5
mln to 2.2).
For a batch multiplication of 5 points it increases the gate count in
ultra from ~72k to ~78k

---------

Co-authored-by: Rumata888 <isennovskiy@gmail.com>