feat!: TXE (#6985)
This PR adds the TXE (Testing eXecution Environment, or Tricky PXE) as a
minimalistic oracle resolver that mimics the behavior of the Aztec
protocol in order to write smart contract tests, much like Foundry for
Solidity contracts.

Together with a `TestEnvironment` added to `aztec-nr`, its aim is to
make possible writing Aztec tests in pure noir resulting in much faster
tests times than running e2e tests.

TXE does NOT test the protocol itself and might let you do very illegal
stuff ATM, but its intent is to provide a faster way to test contract
logic and contract logic only.

A 38 second Token Transfer test:

![Screenshot 2024-06-14 at 13 38
08](https://github.com/AztecProtocol/aztec-packages/assets/5404052/93c38be5-b2fd-4507-9894-e3c93e9bb913)

WARNING: TXE is under heavy development and we are learning how and what
we want to test as we go. It's also in need of a heavy refactor, but we
need this base code in order to get it in the hands of more people and
understand its scope and required capabilities.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>