refactor: add log_hash as input in log emission in private context (#7249)
Updates the `PrivateContext` such that the `emit_raw_note_log` and
`emit_ raw_event_log_with_masked_address` now both are taking the
`log_hash` as a parameter instead of computing it.

This allow us more flexibility when designing note emission wrappers,
meaning that we can have the one we have currently, where the full
encryption and log hash computation is constrained, but it is also
possible to more easily support it being the case that these are instead
unconstrained, but still keeping the same interface.