feat: MSM sorting (#7351)
Implements logic for reducing MSM inputs by first summing points which
share a scalar. See description in `sorted_msm.hpp` for details of the
approach. This is useful when the `scalars` input for an MSM contains
many repeated values because point addition is much cheaper than scalar
multiplication. For example, when using the structured trace, the
permutation grand product polynomial is non-zero but constant over
'dead' regions between blocks.

Note: For now this work is not integrated into the proving systems. That
will be done in a follow on. At that time, considerations will be made
for memory consumption (e.g. perhaps memory allocated in
pippenger_runtime_state can be repurposed for the MSM sorting logic
etc.)

---------

Co-authored-by: zac-williamson <blorktronics@gmail.com>