feat: Run block-proving jobs in parallel by forking world-state (#7655)
# Goal

We want to be able to kick off more than one `prove(blocknumber)` job in
the prover-node at the same time.

We currently cannot do it because the prover-node has a single
world-state, and building a proof modifies world-state. In particular,
preparing the inputs for base rollups modifies state trees, and
finalising the block modifies the archive tree.

# Why?

This'll be needed for the proving integration contest, in case we
generate blocks faster than the time it takes to prove them.

It may still be useful once we move to epoch proving and
sequencer-prover coordination, since the same prover could be picked for
generating proofs for two consecutive epochs.

# How?

## The easy way

Easiest approach is to keep everything as-is today, and clone the world
state before kicking off a job. Eventually, once we implement [Phil's
world-state](https://github.com/AztecProtocol/engineering-designs/pull/9),
we can use the writeable world-state snapshots for this. **This is what
we're doing on this PR.**

## The not-so-easy way

Another approach is to decouple input-generation from proving. Today the
prover-orchestrator is responsible for computing all inputs, but this is
not strictly needed. We can have one component that generates all
inputs, modifies world-state, and outputs a graph of proving jobs (as
Mitch commented
[here](https://aztecprotocol.slack.com/archives/C04BTJAA694/p1722195399887399?thread_ts=1722195378.794149&cid=C04BTJAA694)).
And then another component that orchestrates the execution of proving
jobs exclusively based on their dependencies.

Note that this new component would be a good fit for generating a block
in the sequencer, which today runs an orchestrator without proving
enabled to get to a block header.

It's unclear whether this component should run everything serially (like
[the old block
builder](https://aztecprotocol.slack.com/archives/C04BTJAA694/p1722195399887399?thread_ts=1722195378.794149&cid=C04BTJAA694)
did), or it makes more sense to fan out circuit simulation jobs to
workers (like the proving orchestrator can do now).