chore(bb): prereq work for polynomial mem optimization (#7949)
This PR: Splitting the implementation of plonk and honk into different
polynomial classes.

Greater work: We will be moving to 'polynomials with more structure' ->
aka tracking a size and a virtual size* where 'size' is actually
allocated memory, and virtual size is conceptually all 0's. This should
save a good deal of memory.

* I would call virtual size one more than the degree of the polynomial,
but it's actually a bit weird to call it that when [size, virtual size)
indices are all 0 coefficients. It's more like one more than the maximum
degree of the polynomial, and only really used as an upper bound for
loops or debug bounds checks.