chore: Configuration for bot on various networks (#8063)
This PR introduces more configuration to the bot to enable the flushing
of setup transactions.

It also configures the networks as follows -

Alphanet

- Don't flush setup transactions
- Max pending txs == 1
- Sequencer min txs per block = 1

Devnet

- Don't flush setup transactions
- Max pending txs == 1
- Sequencer min txs per block = 1

Provernet

- Do flush setup transactions
- Max pending txs == 20
- Sequencer min txs per block = 4

The aim is to ensure that Alphanet and Devnet continue to advance the
chain if there are no other txs available. Provernet builds blocks of 4
txs and aims to keep the pool around 20 txs deep.

---------

Co-authored-by: Santiago Palladino <santiago@aztecprotocol.com>