chore(ci): fix bb publishing (#8486)
bb releases are broken again due to using deprecated actions.

I've also added the `issues` permission to this workflow as the action
to raise issues when nightly builds fail isn't working although we don't
need this in the noir org for some reason (maybe there's some org setup
which differs).