chore: remove key rotation from the key store (#8645)
Follow up to #8613 - the key store still had some lingering code from
back when it supported key rotation - each entry could hold multiple
keys. We no longer do this (i.e. we never append anything to an entry,
we only set it with a single key), so instead of checking if the entry
includes the hash we're looking for, we can simply check if it's the
correct (only) one.

Similarly when looking for the public key and secret key we no longer
need to iterate over the multiple entries since there's a single one.