feat: Prover node sends quotes on new epochs (#8864)
Modifies the prover node so it no longer automatically attempts to prove
unproven blocks, but instead sends quotes to the proof coordinator. Also
monitors for claims, and if one that matches the prover address is
posted, only then starts a new epoch proving job.

Other changes include:
- Fixes `assumeProvenThroughBlockNumber` in Rollup so it fakes a claim
for the epoch, otherwise `nextEpochToProve` reports an old epoch and
confuses the sequencer.
- Deletes the e2e-prover-node test, replacing it with the
`e2e_prover/full`. Adds the `e2e_prover/full` with `FAKE_PROOFS` set to
run on every PR. Deletes the `e2e_prover/with_padding` which was not
testing anything (didn't even wait for the proofs).
- Adds new CheatCodes class for playing around with the Rollup contract
itself.
- Adds new methods to the archiver for returning the current L2 slot and
epoch, for retrieving blocks based on epoch number, and for determining
whether an epoch is complete.
- Changes the default behaviour of
`SnapshotManager.publicDeployAccounts` so it doesn't wait for accounts
to be proven.
- Adds new env vars `PROVER_NODE_POLLING_INTERVAL_MS`,
`QUOTE_PROVIDER_BASIS_POINT_FEE`, and `QUOTE_PROVIDER_BOND_AMOUNT`.
- Deletes the `aztec-node-prover-coordination` wrapper and instead just
works with an `AztecNode`.
- Adds methods to the L1 publisher to get the epoch to prove and current
claim. These methods should not go in there, or we should rename it to
something more general than a publisher.

Fix #8684
Fix #8683