refactor: partial notes macros (#8993)
I realized that we need to provide a way to automatically encode the unencrypted log as well. For this reason I decided to refactor the partial notes macro such that we have a `FinalizationPayload` which takes the note hiding point and the nullable values as args to constructor. Along with this change it made sense to rename `PartialPayload` as `SetupPayload`.

I also went ahead and polished all the note macro code.