feat: sync tagged logs (#9595)
Closes: https://github.com/AztecProtocol/aztec-packages/issues/9380

Implements the logic to sync tagged note logs from the node, taking into
account the indices of "last seen" ones.

Unfortunately, the approach of modularizing it as oracles and putting
the logic in the contract itself was met with limitations in noir
(namely, nested slices) and even the code is ready and close to becoming
separate oracles, it cannot be done (yet! or at least without major
caveats and performance implications)