feat: add miscellaneous block to handle structure trace overflow (#9733)
Adds an `overflow` block to the set of execution trace blocks which can
store arbitrary gate types that have "overflowed" the capacity of their
corresponding designated gate type block. This allows for accumulation
of app circuits which do not fit into a given structured trace
configuration.

It works as follows: when constructing a circuit, there is no bound on
the size of each gate block. When a proving key is constructed, we check
the capacity of each block and move any overflow into the overflow
block. In all cases, the polynomials have the prescribed trace structure
at the beginning plus a possible overflow block at the end.

---------

Co-authored-by: iakovenkos <sergey.s.yakovenko@gmail.com>