feat: cli wallet improvements (#10425)
`cli-wallet` now supports an integrated PXE, so it can connect to a
real-er network directly. A new command was also added to register a
recipient so the new note discovery and syncing process is supported.

Also found a bug in the synchronizer that prevented PXE from recovering
when starting from a pre-seeded db (it always set its own header to the
one of the genesis block, regardless of stored state)