feat: PXE sync on demand (#10613)
Closes: https://github.com/AztecProtocol/aztec-packages/issues/10296

Removed the running promise of the synchronizer, in favor of manually
triggering syncs.

Right now, this happens before simulation (tx and unconstrained)

The concept of synchronizer is kept, since we still have to contact the
node and perform some cleanups in case of a reorg.

Should deflake `e2e_2_pxes` and finally means the end of PXE as a
framework, in favor of more library-like approach.