chore: Avoid getport race conditions when starting anvil (#11077)
Our test suite uses `startAnvil` all over the place to start an anvil
instance in the background. This was using `get-port`, which was subject
to race conditions when checking for port availability if we ran
multiple tests in parallel.

This PR removes `get-port` and instead uses port zero to ask the OS to
allocate a port for us, which should be free of race conditions (see
[this SO answer](https://unix.stackexchange.com/a/55918)).