feat: Gemini PolynomialBatcher (#11398)
Introduces new class in Gemini called `PolynomialBatcher` that is
responsible for storing references to the polynomials to be batched and
actually computing the various batched polynomials required in Gemini.
This serves two purposes: (1) it separates the polynomial batching logic
from the "gemini" logic (i.e. constructing univariate claims from a
multilinear poly via gemini-style "folding"), and (2) it facilitates the
input of different types of polynomial sets into Gemini. This latter
point is needed for the new merge protocol which will add yet another
type of polynomial into the mix: `to_be_k_shifted_polynomials`, which
the ever-expanding gemini interface with defaulted inputs would not
easily support.

Note: Currently the `PolynomialBatcher` only handles the `unshifted` and
`to_be_shifted` polynomials. It would be natural to include the
concatenation polynomials as well. I held off for now due to expected
changes to the way concatenations are handled.

Note 2: Upcoming follow ons will introduce a similar mechanism for the
analogous verifier logic.