chore: turn on masking in ultra and mega zk + oink clean-up (#11693)
We have a mechanism to mask witness commitments and evaluations in ZK
Flavors, in this PR the masking is enabled.

It is also a precursor to short scalars in UH ZK Recursive verifier, as
it eliminates the edge cases from `bn254_endo_batch_mul`.

Cleaned up oink prover using a newly introduced `commit_with_type`
method